Split DNSOutgoing/DNSIncoming/DNSMessage into zeroconf._protocol
